requirements.txt
PyQt5>=5.15
requests>=2.28
aiohttp>=3.8
beautifulsoup4>=4.11
transformers>=4.25
playwright>=1.32
PyYAML>=6.0
win10toast>=0.9
weasyprint>=53.3
torch>=1.13.1
